Simplify the serializer
